Add companyid to orgmember (#664)
