+++ rally-domain/src/integration-test/groovy/yakworks/rally/orgs/OrgRepoTests.groovy
@@ -1,6 +1,7 @@
 package yakworks.rally.orgs
 
 import yakworks.commons.map.Maps
+import yakworks.rally.orgs.model.Company
 import yakworks.rally.orgs.model.Location
 import yakworks.rally.orgs.model.OrgCalc
 import yakworks.rally.orgs.model.OrgFlex
@@ -265,6 +266,8 @@ class OrgRepoTests extends Specification implements DomainIntTest {
         result.member != null
         result.member.division.id == division.id
         result.member.business.id == division.member.business.id
+        result.member.company
+        result.member.companyId == Company.DEFAULT_COMPANY_ID
 
         when:
         Org otherBusiness = Org.of("b2", "b2", OrgType.Business).persist([flush: true])
